backend/requirements.txt
Django==3.2.16
django-filter==23.5
djangorestframework==3.12.4
djoser==2.1.0
Pillow==9.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
django-cors-headers==3.13.0
python-dotenv==1.0.1